restructured new updates additions, removed need for running monitor before making api calls, supported changes to stream id
